modificado models, routes e  requisições

+++ backend/routes/comments.js
@@ -1,38 +1,62 @@
 const express = require('express');
 const router = express.Router();
-const Comment = require('../models/Comments');
-const Post = require('../models/Post');
+const authenticateToken = require('../middlewares/authenticateToken');
+const Comment = require('../models/Comment');
 
-// Rota para criar um novo comentário
-router.post('/:postId/comments', async (req, res) => {
-    const postId = req.params.postId;
-    const { text } = req.body;
+/// Rota para obter todos os comentários de um post
+router.get('/:postId', async (req, res) => {
     try {
-        const newComment = await Comment.createComment(postId, text);
-        res.json(newComment);
+      const postId = req.params.postId;
+      const comments = await Comment.getAllComments(postId);
+      res.json(comments);
     } catch (error) {
-        console.error(error);
-        res.status(500).json({ error: 'Erro ao criar o comentário.' });
+      console.error(error);
+      res.status(500).json({ error: 'Erro ao obter os comentários.' });
     }
-});
-
-// Rota para obter todos os comentários de um post específico
-router.get('/:postId/comments', async (req, res) => {
-    const postId = req.params.postId;
+  });
+  
+  // Rota para criar um comentário (autenticada)
+  router.post('/create', authenticateToken, async (req, res) => {
     try {
-        const comments = await Comment.getAllComments(postId);
-        res.json(comments);
+      const userId = req.user.id;
+      const { postId, content } = req.body;
+  
+      const newComment = await Comment.createComment(userId, postId, content);
+  
+      res.status(201).json({ message: 'Comentário criado com sucesso.', comment: newComment });
+    } catch (error) {
+      console.error(error);
+      res.status(500).json({ error: 'Erro ao criar o comentário.' });
+    }
+  });
+
+// Rota para editar um comentário existente
+router.put('/:commentId', authenticateToken, async (req, res) => {
+    const commentId = req.params.commentId;
+    const { content } = req.body;
+    const userId = req.user.userId;
+
+    try {
+        const comment = await Comment.getCommentById(commentId);
+        if (comment.userId !== userId) {
+            return res.status(403).json({ error: 'Usuário não autorizado para editar este comentário.' });
+        }
+
+        const updatedComment = await Comment.editComment(commentId, content);
+        res.json(updatedComment);
     } catch (error) {
         console.error(error);
-        res.status(500).json({ error: 'Erro ao obter os comentários.' });
+        res.status(500).json({ error: 'Erro ao editar o comentário.' });
     }
 });
 
 // Rota para curtir um comentário
-router.put('/comments/:commentId/like', async (req, res) => {
+router.put('/:commentId/like', authenticateToken, async (req, res) => {
     const commentId = req.params.commentId;
+    const userId = req.user.userId;
+
     try {
-        await Comment.likeComment(commentId);
+        await Comment.likeComment(userId, commentId);
         res.json({ message: 'Comentário curtido com sucesso!' });
     } catch (error) {
         console.error(error);
@@ -40,11 +64,31 @@ router.put('/comments/:commentId/like', async (req, res) => {
     }
 });
 
-// Rota para excluir um comentário de um post específico
-router.delete('/:postId/comments/:commentId', async (req, res) => {
-    const postId = req.params.postId;
+// Rota para descurtir um comentário
+router.put('/:commentId/unlike', authenticateToken, async (req, res) => {
     const commentId = req.params.commentId;
+    const userId = req.user.userId;
+
     try {
+        await Comment.unlikeComment(userId, commentId);
+        res.json({ message: 'Comentário descurtido com sucesso!' });
+    } catch (error) {
+        console.error(error);
+        res.status(500).json({ error: 'Erro ao descurtir o comentário.' });
+    }
+});
+
+// Rota para excluir um comentário existente
+router.delete('/:commentId', authenticateToken, async (req, res) => {
+    const commentId = req.params.commentId;
+    const userId = req.user.userId;
+
+    try {
+        const comment = await Comment.getCommentById(commentId);
+        if (comment.userId !== userId) {
+            return res.status(403).json({ error: 'Usuário não autorizado para excluir este comentário.' });
+        }
+
         await Comment.deleteComment(commentId);
         res.json({ message: 'Comentário excluído com sucesso!' });
     } catch (error) {
@@ -53,20 +97,6 @@ router.delete('/:postId/comments/:commentId', async (req, res) => {
     }
 });
 
-// Rota para excluir todos os comentários de um post específico antes de excluir o post
-router.delete('/:postId/comments/all', async (req, res) => {
-    const postId = req.params.postId;
-    try {
-        await Comment.deleteAllComments(postId);
-
-        await Post.deletePost(postId);
-
-        res.json({ message: 'Post e seus comentários foram excluídos com sucesso!' });
-    } catch (error) {
-        console.error(error);
-        res.status(500).json({ error: 'Erro ao excluir o post e seus comentários.' });
-    }
-});
 
 
 module.exports = router;
